refactor: encapsulate backend calls
Logic specific to different backends lives in backend.
Lib should be a relatively thin wrapper which just
picks the backend.

The actual logic of the backend should not leak to lib
beyond a simple function call.
